+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/streams/darttoplatform/ImageByteStreamHandler.kt
@@ -14,7 +14,6 @@ import deckers.thibault.aves.glide.AvesAppGlideModule
 import deckers.thibault.aves.model.EntryFields
 import deckers.thibault.aves.utils.BitmapUtils
 import deckers.thibault.aves.utils.BitmapUtils.applyExifOrientation
-import deckers.thibault.aves.utils.ContextUtils.devicePixelRatio
 import deckers.thibault.aves.utils.LogUtils
 import deckers.thibault.aves.utils.MimeTypes
 import deckers.thibault.aves.utils.MimeTypes.handleEncodedBytesInFlutter
@@ -27,13 +26,11 @@ import kotlinx.coroutines.withContext
 import java.io.ByteArrayInputStream
 import java.io.InputStream
 import java.util.Date
-import kotlin.math.roundToInt
 
 class ImageByteStreamHandler(private val context: Context, private val arguments: Any?) : BaseStreamHandler(), ByteSink {
     private var op: String? = null
     private var decoded: Boolean = false
     private val regionFetcher = RegionFetcher(context)
-    private val density = context.devicePixelRatio()
 
     init {
         if (arguments is Map<*, *>) {
@@ -231,7 +228,7 @@ class ImageByteStreamHandler(private val context: Context, private val arguments
             return
         }
 
-        val uri = arguments[EntryFields.URI] as String?
+        val uri = (arguments[EntryFields.URI] as String?)?.toUri()
         val pageId = arguments["pageId"] as Int?
         val mimeType = arguments[EntryFields.MIME_TYPE] as String?
         val dateModifiedMillis = (arguments[EntryFields.DATE_MODIFIED_MILLIS] as Number?)?.toLong()
@@ -239,10 +236,8 @@ class ImageByteStreamHandler(private val context: Context, private val arguments
         val isFlipped = arguments[EntryFields.IS_FLIPPED] as Boolean?
         val widthDip = (arguments["widthDip"] as Number?)?.toDouble()
         val heightDip = (arguments["heightDip"] as Number?)?.toDouble()
-        val defaultSizeDip = (arguments["defaultSizeDip"] as Number?)?.toDouble()
-        val quality = arguments["quality"] as Int?
 
-        if (uri == null || mimeType == null || rotationDegrees == null || isFlipped == null || widthDip == null || heightDip == null || defaultSizeDip == null || quality == null) {
+        if (uri == null || mimeType == null || rotationDegrees == null || isFlipped == null || widthDip == null || heightDip == null) {
             error("getThumbnail-args", "missing arguments", null)
             return
         }
@@ -257,10 +252,8 @@ class ImageByteStreamHandler(private val context: Context, private val arguments
             dateModifiedMillis = dateModifiedMillis ?: (Date().time),
             rotationDegrees = rotationDegrees,
             isFlipped = isFlipped,
-            width = (widthDip * density).roundToInt(),
-            height = (heightDip * density).roundToInt(),
-            defaultSize = (defaultSizeDip * density).roundToInt(),
-            quality = quality,
+            widthDip = widthDip,
+            heightDip = heightDip,
             result = this,
         ).fetch()
     }